battle handler
